DOing stuff

+++ src/Vector3.java
@@ -1,7 +1,7 @@
 public class Vector3 {
-    public int x, y, z;
+    public double x, y, z;
 
-    public Vector3(int x, int y, int z) {
+    public Vector3(double x, double y, double z) {
         this.x = x;
         this.y = y;
         this.z = z;
